Replace bitflags with bitfield
While making this change it was realised that there was a bug in that
the bitflags created weren't actually being used. The bytecode_options
field had been set as being of type u8 and not ByteCodeOptions.

The reason for switching between the two crates was because later
definitions of the bytecode format require bitfield. The bitflags crate
only supports bitflags while the bitfield crate can handle both
situations.

Signed-off-by: Victor Omondi <victor@slick.co.ke>