Refactor main.rs
The parsers' return type was too verbose. Since the first element of
the tuple remains the same, it made sense to create a type alias to
reduce code noise.

The multi_parser function serves to prevent repetition since a lot of
the bytecode segments share a similar pattern.

Also fixed a bug in bytecode_format.rs where the fields for the
StringKind struct were not public

Signed-off-by: Victor Omondi <victor@slick.co.ke>