Refactor Identifier module
The initial version of the IR ended up being too complicated. The
amount of nested structured would make control flow analysis very
challenging. The first step in simplifying things is realising that
the opcodes deal primarily with registers and occasionaly literal
values. With this in mind, the Identifier type had to be split.

The StringIndex is now considered a Literal while the Register type
shall fulfill the role of "Identifier". As a result the module had
to be renamed so that it better communicates what it contains.

All literals now derive from Copy and Clone as they are relatively
small structures hence shouldn't be expensive to copy.

Signed-off-by: Victor Omondi <victor@slick.co.ke>