Refator object module
The goal of these changes was to develop an IR that enables a few
structures to emulate various Hermes opcodes. This would greatly
simplify the analysis and code generation processes.

The NewObjectExpression struct caters for all NewObject* opcodes.
Similarly, the NewArrayExpression deals with NewArray* opcodes.

EnvExpression will be used on the opcodes that deal with the
Hermes Environment used by closures.

ObjectExpression will represent the opcodes for getting and setting
property values in Objects. This includes array values.

Signed-off-by: Victor Omondi <victor@slick.co.ke>