Implement function call IR
The builtins module contains the builtin methods supported by
Hermes. This structure may differ from one version of the VM to
another. Giving the enum variants values isn't necessary and is
done just as a way of documenting the total number of builtin
methods.

The FrameCall struct represents function calls that retrieve their
arguments from the end of the current frame. This struct is intended
to be converted to a CallExpression just before any optimizations are
done. This is because CallExpression is self contained thus making it
easier to work with.

Signed-off-by: Victor Omondi <victor@slick.co.ke>