Implement custom Nom error type
The Error types provided by Nom did not satisfy the project's
requirement to have contextual errors. Also, the input bytes
is not needed as part of the Error type. The ParserError type
intends to fulfill these needs.

In order to be compatible with Nom parsers, it implements Nom's
ParseError trait. Implementing the ContextError trait, allows
parsers to use the context function to provide more context when
an Error occurs.

Signed-off-by: Victor Omondi <victor@slick.co.ke>