Add convertion from String to ParserError
This will cater for the scenarios, particularly when implementing
TryFrom, where the error is a String. The TryFrom trait will be
implemented in most IR types to try and convert from an Opcode variant,
and thus the error will most likely appear during parsing.

The error for "TryFrom<u8> for Opcode" was changed not only as part of
the above reasoning, but also to provide more context on the error.

Signed-off-by: Victor Omondi <victor@slick.co.ke>