Refactor TryFrom error type
It was realised that having the error type as String would be an
issue since as stated in commit 52c57e9, the error would occur during
parsing. Therefore there would need to be an explicit conversion from
string to nom::Err<> type.

To simplify things, the types implementing TryFrom now return a
ParserError. Then, by implementing From<ParserError> for
nom::Err<ParserError> an implicit conversion can happen when using the
?.

Signed-off-by: Victor Omondi <victor@slick.co.ke>