Refactor Register
In commit: 5b69bf, the From trait was implemented for the variants
of Register. This was meant to be the way to initialise them. Looking
back, this was very unnecessary. The variants could be initalised
directly without the need for the From trait for each.

These changes remove the From traits.

Signed-off-by: Victor Omondi <victor@slick.co.ke>