Simplify magic_parser's Error
Having already implemented the From<ParserError> trait for nom::Err,
explicitly initialising one here was too verbose. The implicit
conversion using the ? is much better.

The or was substituted with or_else since the latter function does
lazy evaluation hence preventing the conversion from being done up
until it is needed. Otherwise, the use of or together with the ?
would result in an Error even in valid circumstances.

Signed-off-by: Victor Omondi <victor@slick.co.ke>