Refactor Environment related types
The StoreToEnvironment opcodes do not fit the requirements of being
an Expression Statement. This is because they don't have a register
that is the result of an expression. This is what led to creating a
new Statement variant to cater for it.

The EnvExpression type was missing a variable that would store the
environment register.

As for the Environment variant in Object, the variable id wasn't
meant to be of type EnvIndex but rather of type u8

Signed-off-by: Victor Omondi <victor@slick.co.ke>