Refactor Return variant in Statement
The logic behind it containing an Option<Expression> was because
a function could return nothing. But since the HBC explicitly dictates
a register to be the argument for the return statement, then this
change is necessary.

When generating the javascript code, the analysis can be done to determine
which kind of expression the register represents.

Signed-off-by: Victor Omondi <victor@slick.co.ke>